fix read_file not setting reader
