+++ tests/v1/test_stac_io.py
@@ -10,34 +10,10 @@ import pytest
 
 import pystac
 import pystac.errors
-from pystac.reader import StandardLibraryReader, Reader
-from pystac.writer import StandardLibraryWriter, Writer
 from pystac.stac_io import DefaultStacIO, DuplicateKeyReportingMixin, StacIO
 
 from .utils import TestCases
-
-
-class MockReader(Reader):
-    def __init__(self) -> None:
-        self.calls: list[str] = []
-        self.reader = StandardLibraryReader()
-
-    def get_json(self, href: str | Path) -> dict[str, Any]:
-        self.calls.append(str(href))
-        return self.reader.get_json(href)
-
-
-class MockWriter(Writer):
-    def __init__(self) -> None:
-        self.calls: list[tuple[dict[str, Any], str | Path]] = []
-        self.writer = StandardLibraryWriter()
-
-    def put_json(self, data: dict[str, Any], href: str | Path) -> None:
-        self.calls.append((data, href))
-        self.writer.put_json(data, href)
-
-    def delete(self, href: str | Path) -> None:
-        self.writer.delete(href)
+from tests.utils.mock_io import MockReader, MockWriter
 
 
 def test_read_write_collection() -> None:
